Breaking down createWiki.ts
