+++ src/utils/CreateWikiUtils/createWiki.ts
@@ -1,62 +1,23 @@
 import config from '@/config'
 import axios from 'axios'
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 import { POST_IMG } from '@/services/wikis/queries'
 import {
   Wiki,
   EditorContentOverride,
-  whiteListedDomains,
-  EditSpecificMetaIds,
-  whiteListedLinkNames,
-  CreateNewWikiSlug,
   LinkedWikiKey,
   LinkedWikis,
 } from '@everipedia/iq-utils'
 import { useAppDispatch } from '@/store/hook'
 import { createContext } from '@chakra-ui/react-utils'
-import { submitVerifiableSignature } from '@/utils/WalletUtils/postSignature'
-import {
-  useAccount,
-  useFeeData,
-  useSignTypedData,
-  useWaitForTransaction,
-} from 'wagmi'
 import { NextRouter } from 'next/router'
 import { skipToken } from '@reduxjs/toolkit/dist/query'
-import { getWiki, useGetWikiQuery } from '@/services/wikis'
-import {
-  getDraftFromLocalStorage,
-  removeDraftFromLocalStorage,
-} from '@/store/slices/wiki.slice'
+import { useGetWikiQuery } from '@/services/wikis'
+import { getDraftFromLocalStorage } from '@/store/slices/wiki.slice'
 import { useToast } from '@chakra-ui/toast'
-import { store } from '@/store/store'
-import { Dict } from '@chakra-ui/utils'
 import { useGetWikiByActivityIdQuery } from '@/services/activities'
 import { WikiImageObjectProps } from '@/types/CreateWikiType'
-import { logEvent } from '../googleAnalytics'
-import { getDeadline } from '../DataTransform/getDeadline'
-import { isValidUrl } from '../textUtils'
-import {
-  defaultErrorMessage,
-  initialEditorValue,
-  initialMsg,
-  successMessage,
-} from './createWikiMessages'
-
-export const domain = {
-  name: 'EP',
-  version: '1',
-  chainId: Number(config.chainId),
-  verifyingContract: config.wikiContractAddress,
-}
-
-export const types = {
-  SignedPost: [
-    { name: 'ipfs', type: 'string' },
-    { name: 'user', type: 'address' },
-    { name: 'deadline', type: 'uint256' },
-  ],
-}
+import { initialEditorValue, initialMsg } from './createWikiMessages'
 
 export const MINIMUM_WORDS = 100
 
@@ -120,175 +81,6 @@ export const useCreateWikiEffects = () => {
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [dispatch, slug])
-}
-
-export const useGetSignedHash = () => {
-  const {
-    setWikiHash,
-    wikiHash,
-    setMsg,
-    setIsLoading,
-    setTxHash,
-    setActiveStep,
-    txHash,
-    setCommitMessage,
-    dispatch,
-  } = useCreateWikiContext()
-
-  const { address: userAddress, isConnected: isUserConnected } = useAccount()
-  const deadline = useRef(0)
-
-  const {
-    data: signData,
-    error: signError,
-    isLoading: signing,
-    signTypedDataAsync,
-  } = useSignTypedData()
-
-  const { refetch } = useWaitForTransaction({ hash: txHash })
-  const { data: feeData } = useFeeData({
-    formatUnits: 'gwei',
-  })
-  const gasPrice = useMemo(
-    () => parseFloat(feeData?.formatted.gasPrice || '0'),
-    [feeData],
-  )
-
-  const saveHashInTheBlockchain = async (ipfs: string) => {
-    deadline.current = getDeadline()
-    setWikiHash(ipfs)
-    signTypedDataAsync({
-      domain,
-      types,
-      value: {
-        ipfs,
-        user: userAddress,
-        deadline: deadline.current,
-      },
-    })
-      .then(response => {
-        if (response) {
-          setActiveStep(1)
-        } else {
-          setIsLoading('error')
-          setMsg(defaultErrorMessage)
-        }
-      })
-      .catch(err => {
-        setIsLoading('error')
-        setMsg(err.message || defaultErrorMessage)
-        logEvent({
-          action: 'SUBMIT_WIKI_ERROR',
-          label: err.message,
-          category: 'wiki_error',
-          value: 1,
-        })
-      })
-  }
-
-  const verifyTrxHash = useCallback(
-    async () => {
-      let timePassed = 0
-      const timer = setInterval(() => {
-        if (timePassed >= 60 * 1000 && gasPrice > 250) {
-          setMsg(`A little congestion on the polygon chain is causing a delay in the 
-          creation of your wiki.This would be resolved in a little while.`)
-        }
-        try {
-          const checkTrx = async () => {
-            const trx = await refetch()
-            if (trx.error || trx.data?.status === 0) {
-              setIsLoading('error')
-              setMsg(defaultErrorMessage)
-              logEvent({
-                action: 'SUBMIT_WIKI_ERROR',
-                label: 'TRANSACTION_VERIFICATION_ERROR',
-                category: 'wiki_error',
-                value: 1,
-              })
-              clearInterval(timer)
-            }
-            if (
-              trx &&
-              trx.data &&
-              trx.data.status === 1 &&
-              trx.data.confirmations > 1
-            ) {
-              setIsLoading(undefined)
-              setActiveStep(3)
-              setMsg(successMessage)
-              // clear all edit based metadata from redux state
-              Object.values(EditSpecificMetaIds).forEach(id => {
-                dispatch({
-                  type: 'wiki/updateMetadata',
-                  payload: {
-                    id,
-                    value: '',
-                  },
-                })
-              })
-              setCommitMessage('')
-              removeDraftFromLocalStorage()
-              clearInterval(timer)
-            }
-          }
-          checkTrx()
-        } catch (err) {
-          const errorObject = err as Dict
-          setIsLoading('error')
-          setMsg(defaultErrorMessage)
-          logEvent({
-            action: 'SUBMIT_WIKI_ERROR',
-            label: errorObject.message,
-            category: 'wiki_error',
-            value: 1,
-          })
-          clearInterval(timer)
-        }
-        timePassed += 3000
-      }, 3000)
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [refetch],
-  )
-
-  useEffect(() => {
-    const getSignedTxHash = async () => {
-      if (signData && wikiHash && isUserConnected && userAddress) {
-        if (signError) {
-          setMsg(defaultErrorMessage)
-          setIsLoading('error')
-          return
-        }
-        try {
-          const hash = await submitVerifiableSignature(
-            signData,
-            wikiHash,
-            userAddress,
-            deadline.current,
-          )
-          if (hash) {
-            setTxHash(hash)
-            setActiveStep(2)
-          }
-        } catch (err) {
-          const errorObject = err as Dict
-          setIsLoading('error')
-          setMsg(errorObject.response.errors[0].extensions.exception.reason)
-          logEvent({
-            action: 'SUBMIT_WIKI_ERROR',
-            label: errorObject.response.errors[0].extensions.exception.reason,
-            category: 'wiki_error',
-            value: 1,
-          })
-        }
-      }
-    }
-    getSignedTxHash()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [signData, signError])
-
-  return { signing, saveHashInTheBlockchain, verifyTrxHash, txHash }
 }
 
 export const useCreateWikiState = (router: NextRouter) => {
@@ -385,48 +177,4 @@ export const useCreateWikiState = (router: NextRouter) => {
     txError,
     setTxError,
   }
-}
-
-export const isVerifiedContentLinks = (content: string) => {
-  const markdownLinks = content.match(/\[(.*?)\]\((.*?)\)/g)
-  let isValid = true
-  markdownLinks?.every(link => {
-    const linkMatch = link.match(/\[(.*?)\]\((.*?)\)/)
-    const text = linkMatch?.[1]
-    const url = linkMatch?.[2]
-
-    if (
-      text &&
-      url &&
-      whiteListedLinkNames.includes(text) &&
-      !isValidUrl(url)
-    ) {
-      isValid = true
-      return true
-    }
-
-    if (url && url.charAt(0) !== '#') {
-      const validURLRecognizer = new RegExp(
-        `^https?://(www\\.)?(${whiteListedDomains.join('|')})`,
-      )
-      isValid = validURLRecognizer.test(url)
-      return isValid
-    }
-    return true
-  })
-  return isValid
-}
-
-export const isWikiExists = async (
-  slug: string,
-  setExistingWikiData: (data: Wiki) => void,
-) => {
-  if (slug === CreateNewWikiSlug) return false
-  const { data, isError } = await store.dispatch(getWiki.initiate(slug))
-  if (isError) return false
-  if (data) {
-    setExistingWikiData(data)
-    return true
-  }
-  return false
 }
